Classes and add del phone
Classese moved in diffrent files. Add remover phone number from contact
