+++ home_work/address_book.py
@@ -1,82 +1,6 @@
 from collections import UserDict
 from datetime import datetime, timedelta
-
-
-class Field:
-
-    def __init__(self, value: str):
-        self.value = value
-
-    def __str__(self):
-        return str(self.value)
-
-
-class Name(Field):
-    pass
-
-
-class Phone(Field):
-
-    def __init__(self, value: str):
-
-        if len(value) != 10 or not value.isdigit():
-            raise ValueError()
-
-        super().__init__(value)
-
-
-class Birthday(Field):
-
-    def __init__(self, value: str):
-        try:
-            self.value = datetime.strptime(value, "%d.%m.%Y").date()
-        except ValueError:
-            raise ValueError(
-                f'Invalid date format. Use DD.MM.YYYY\n'f"Or day {value} is out of range for month")
-
-
-class Record:
-
-    def __init__(self, name: str):
-        self.name = Name(name)
-        self.phones = []
-        self.birthday = None
-
-    def __str__(self):
-
-        birthday_str = f", birthday: {self.birthday.value.strftime('%d.%m.%Y')}" if self.birthday else ""
-        return f"Contact name: {self.name.value.title()}, phones: {'; '.join(p.value for p in self.phones)}{birthday_str}"
-
-    def add_phone(self, phone: str):
-        self.phones.append(Phone(phone))
-
-    def add_birthday(self, birth_date: str):
-
-        self.birthday = Birthday(birth_date)
-        return f"{self.birthday.value.strftime('%d.%m.%Y')} added"
-
-    def remove_phone(self, rem_phone: str):
-
-        self.phones = [
-            phone for phone in self.phones if phone.value != rem_phone]
-
-    def edit_phone(self, old_phone: str, new_phone: str):
-
-        for phone in self.phones:
-            if phone.value == old_phone:
-                phone.value = Phone(new_phone).value
-
-                return f"Номер {old_phone} змінено на номер {new_phone}"
-
-        raise ValueError(f"Номер {old_phone} не знайдено")
-
-    def find_phone(self, f_phone: str):
-
-        for phone in self.phones:
-            if phone.value == f_phone:
-                return phone
-
-        return None
+from record import Record
 
 
 class AddressBook(UserDict):
