+++ home_work/field.py
@@ -1,0 +1,7 @@
+class Field:
+
+    def __init__(self, value: str):
+        self.value = value
+
+    def __str__(self):
+        return str(self.value)
