+++ home_work/name.py
@@ -1,0 +1,5 @@
+from field import Field
+
+
+class Name(Field):
+    pass
