+++ home_work/phone.py
@@ -1,0 +1,11 @@
+from field import Field
+
+
+class Phone(Field):
+
+    def __init__(self, value: str):
+
+        if len(value) != 10 or not value.isdigit():
+            raise ValueError()
+
+        super().__init__(value)
